UPD: Rename RP to rp
