+++ examples/use_rust/rp2040_split_pio/src/peripheral.rs
@@ -16,7 +16,7 @@ use embassy_rp::{
 use panic_probe as _;
 use rmk::split::{
     peripheral::run_rmk_split_peripheral,
-    RP::uart::{BufferedUart, UartInterruptHandler},
+    rp::uart::{BufferedUart, UartInterruptHandler},
     SPLIT_MESSAGE_MAX_SIZE,
 };
 use static_cell::StaticCell;
