+++ rmk/src/split/mod.rs
@@ -12,7 +12,7 @@ pub mod peripheral;
 #[cfg(not(feature = "_nrf_ble"))]
 pub mod serial;
 #[cfg(feature = "rp2040_pio")]
-pub mod RP;
+pub mod rp;
 
 /// Maximum size of a split message
 pub const SPLIT_MESSAGE_MAX_SIZE: usize = SplitMessage::POSTCARD_MAX_SIZE + 4;
